requirements.txt
flask
openai
python-dotenv
numpy
Pillow==9.5.0
moviepy==1.0.3